requirements.txt
# 딥페이크 탐지 모델 - 로컬 개발 환경
# Python 3.9+ 권장

# 딥러닝 프레임워크
torch==2.7.1
torchvision==0.22.1

# Transformers (ViT 모델)
transformers==4.30.0
datasets

# 컴퓨터 비전
opencv-python==4.10.0.82
Pillow

# 얼굴 검출
dlib

# 데이터 처리
numpy==1.26.4
pandas
scipy==1.11.4
scikit-learn==1.3.2

# 유틸리티
tqdm
pathlib

# 개발용 (선택)
jupyter
ipykernel
matplotlib

